refactor(e2e): tree model
